rename and edit some file.
