+++ app/src/main/java/com/tfandkusu/observeroom/datastore/Member.kt
@@ -3,6 +3,7 @@ package com.tfandkusu.observeroom.datastore
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
+import com.tfandkusu.observeroom.data.Division
 
 /**
  * 社員
@@ -16,4 +17,7 @@ import androidx.room.PrimaryKey
         parentColumns = arrayOf("id"), childColumns = arrayOf("divisionId")
     )]
 )
-data class Member(@PrimaryKey(autoGenerate = true) val id: Long, val name: String, val divisionId: Long)
+data class Member(
+    @PrimaryKey(autoGenerate = true) val id: Long, val name: String,
+    val divisionId: Long
+)
