+++ app/src/main/java/com/tfandkusu/observeroom/datastore/MemberWithDivision.kt
@@ -2,6 +2,7 @@ package com.tfandkusu.observeroom.datastore
 
 import androidx.room.Embedded
 import androidx.room.Relation
+import com.tfandkusu.observeroom.data.Division
 
 data class MemberWithDivision(
     @Embedded val member: Member,
